kocaeli_haber_projesi/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pymongo==4.6.1
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4



















